+++ src/lib/templateStrings.js
@@ -1,5 +1,5 @@
 import utils from "./utils.js";
-import DDBHelper from "./ddb.js";
+import DDBHelper from "./DDBHelper.js";
 import logger from "../logger.js";
 import CompendiumHelper from "./CompendiumHelper.js";
 import { generateAdventureConfig } from "../muncher/adventure.js";